Added more test cases and refactored tests. Added comments.
