+++ src/UnitTests/X509CertificateBasedEncryptor_EncryptStream.cs
@@ -28,4 +28,3 @@ namespace UnitTests
         }
     }
 }
-
